lucene企业级检索实现-前后联调可用版

+++ luceneQuery/src/main/java/com.iris.lucene/QueryMain.java
@@ -11,13 +11,13 @@ public class QueryMain {
         getTotals(luceneQuery);
       }
 
-    private static Integer getTotals(LuceneQuery luceneQuery) {
+    private static Long getTotals(LuceneQuery luceneQuery) {
         Map<String, Object> map = luceneQuery.getTotal();
         Integer total1 = (Integer) map.get("total1");
         Integer total2 = (Integer) map.get("total2");
         Integer total3 = (Integer) map.get("total3");
         Integer total4 = (Integer) map.get("total4");
-        Integer total = (Integer) map.get("total");
+        Long total = (Long) map.get("total");
         List<AuditRecordLuceneNew> resultList = (List<AuditRecordLuceneNew>) map.get("list");
         System.out.println(new Date() + "|【[" + 1 + "]：" + total1 + "】||【[" + 2 + "]：" + total2 + "】||【[" + 3 + "]：" + total3 + "】||【[" + 4 + "]：" + total4 + "】 " + "文件夹总数：" + total);
         resultList.forEach(x->System.out.println(new Gson().toJson(x)));
